backend/requirements.txt
absl-py==2.1.0
addict==2.4.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
einops==0.8.0
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
git-filter-repo==2.45.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.7
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mmcv==2.2.0
mmengine==0.10.4
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
MouseInfo==0.1.3
mpmath==1.3.0
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
openai==1.43.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
pgmpy==0.1.26
pillow==10.4.0
platformdirs==4.2.2
prettytable==3.11.0
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20240210
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.4
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.97
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==306
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
statsmodels==0.14.4
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tokenizers==0.20.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.8.0
Werkzeug==3.0.4
wheel==0.44.0
wikipedia==1.4.0
wrapt==1.16.0
xgboost==2.1.1
yapf==0.40.2
zipp==3.20.1

SQLAlchemy~=2.0.36